+++ notification/src/epilot/models/operations/getnotification.py
@@ -19,9 +19,12 @@ class GetNotificationRequest:
 @dataclasses.dataclass
 class GetNotificationResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     notification_item: Optional[dict[str, Any]] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""