+++ notification/src/epilot/models/operations/gettotalunread.py
@@ -10,9 +10,12 @@ from typing import Optional
 @dataclasses.dataclass
 class GetTotalUnreadResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     get_total_unread_200_text_plain_number: Optional[str] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""