+++ notification/src/epilot/models/operations/markasread.py
@@ -19,7 +19,10 @@ class MarkAsReadRequest:
 @dataclasses.dataclass
 class MarkAsReadResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""